+++ CHANGELOG.md
@@ -10,6 +10,11 @@ This project uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - Added support for `ndarray::Array1` inputs.
 - Added support for `ndarray::Array2` inputs.
 
+### Changed
+
+- Renamed `MatrixLayout::RowWise` and `MatrixLayout::ColumnWise` to
+  `MatrixLayout::RowMajor` and `MatrixLayout::ColumnMajor`.
+
 ## [0.2.0] - 2023-09-02
 
 ### Added