+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 All notable changes to this project will be documented in this file.
 This project uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## [0.3.0] - 2023-09-02
 
 ### Added
 
@@ -31,5 +31,6 @@ This project uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 - 🎉 Initial release.
 
+[0.3.0]: https://github.com/sunsided/timelag-rs/releases/tag/0.3.0
 [0.2.0]: https://github.com/sunsided/timelag-rs/releases/tag/0.2.0
 [0.1.0]: https://github.com/sunsided/timelag-rs/releases/tag/0.1.0